Revert "refactor: move program list styling to `ProgramRunnerInfo` (#134)"
This reverts commit 59736a79c82c0919f95557a699d0f99301c28cd6.
