+++ src/programs/index.ts
@@ -13,26 +13,8 @@ export interface RunningProgram {
 
 // Currently used only for Host, due to a circular dependency
 export interface ProgramRunner {
-  /**
-   * Adds a new program to run.
-   * @param name program name
-   * @param inputs program inputs
-   * @returns the running program data
-   */
-  addRunningProgram(name: string, inputs: string[]): RunningProgram;
-
-  /**
-   * Lists running programs.
-   * @returns the list of running programs
-   */
-  getRunningPrograms(): RunningProgram[];
-
-  /**
-   * Stops a running program.
-   * @param pid running program ID
-   * @returns `true` if the program exists and was stopped, `false` otherwise
-   */
-  removeRunningProgram(pid: Pid): boolean;
+  addRunningProgram(name: string, inputs: string[]): void;
+  removeRunningProgram(pid: Pid): void;
 }
 
 export interface Program {
